requirements.txt
a2a-sdk==0.3.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
attrs==25.4.0
azure-core==1.36.0
azure-identity==1.25.1
boto3==1.40.67
botocore==1.40.67
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
email-validator==2.3.0
fast-agent-mcp==0.3.19
fastapi==0.121.0
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.43.0
google-genai==1.49.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-aiohttp==0.1.9
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jiter==0.11.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mcp==1.19.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
openai==2.7.1
opentelemetry-api==1.38.0
opentelemetry-distro==0.59b0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-anthropic==0.47.5
opentelemetry-instrumentation-google-genai==0.4b0
opentelemetry-instrumentation-mcp==0.47.5
opentelemetry-instrumentation-openai==0.47.5
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-genai==0.2b0
packaging==25.0
platformdirs==4.5.0
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.11.0
pygments==2.19.2
pyjwt==2.10.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-frontmatter==1.1.0
python-multipart==0.0.20
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.49.3
tenacity==9.1.2
tensorzero==2025.11.1
textual==6.5.0
tqdm==4.67.1
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uc-micro-py==1.0.3
urllib3==2.5.0
uuid-utils==0.11.1
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
